pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["daq_connectivity"] 

[project]
name = "daq_connectivity"
version = "0.0.1"
authors = [
  { name="Eduard Almar", email="edu12due@gmail.com" },
]
description = "Compares how smilar is the current data with the historical data"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "matplotlib",
    "numpy",
    "pandas",
    "pyserial",
]

[project.urls]
"Homepage" = "https://github.com/opticsensors/daq_connectivity"
"Bug Tracker" = "https://github.com/opticsensors/daq_connectivity/issues"